feat(condo): DOMA-13345 cards online interaction
