+++ apps/condo/domains/acquiring/schema/AllCardBindingsService.js
@@ -2,14 +2,14 @@
  * Generated by `createservice acquiring.AllCardBindingsService`
  */
 
-const { getLogger } = require('@open-condo/keystone/logging')
-const { GQLCustomSchema, find } = require('@open-condo/keystone/schema')
+const { GQLCustomSchema } = require('@open-condo/keystone/schema')
 
 const access = require('@condo/domains/acquiring/access/AllCardBindingsService')
-const { ACQUIRING_INTEGRATION_ONLINE_PROCESSING_TYPE } = require('@condo/domains/acquiring/constants/integration')
-const { getUserCards } = require('@condo/domains/acquiring/utils/serverSchema/OnlineInteraction')
+const {
+    fetchCardTokens,
+    deduplicateCardTokens,
+} = require('@condo/domains/acquiring/utils/serverSchema/allCardBinding/utils')
 
-const logger = getLogger()
 
 const AllCardBindingsService = new GQLCustomSchema('AllCardBindingsService', {
     types: [
@@ -57,68 +57,10 @@ const AllCardBindingsService = new GQLCustomSchema('AllCardBindingsService', {
                     },
                 } = args
 
-                const acquiringIntegrations = await find('AcquiringIntegration', {
-                    type: ACQUIRING_INTEGRATION_ONLINE_PROCESSING_TYPE,
-                    getUserCardsUrl_not: null,
-                    deleteUserCardUrl_not: null,
-                    deletedAt: null,
-                })
-                
-                const results = await Promise.allSettled(
-                    acquiringIntegrations.map(integration =>
-                        getUserCards(
-                            integration.getUserCardsUrl,
-                            userId,
-                            integration.name,
-                            integration.id,
-                        )
-                    )
-                )
-
-                const cardTokens = []
-
-                for (const result of results) {
-                    if (result.status === 'fulfilled') {
-                        cardTokens.push(...result.value)
-                    } else {
-                        logger.error({
-                            msg: 'Failed to fetch card bindings',
-                            err: result.reason,
-                        })
-                    }
-                }
-
-                const deduplicated = new Map()
-
-                for (const card of cardTokens) {
-                    const key = [
-                        card.id,
-                        card.cardNumber,
-                        card.paymentSystem,
-                        card.expiration,
-                    ].join(':')
-
-                    const existing = deduplicated.get(key)
-
-                    if (!existing) {
-                        deduplicated.set(key, {
-                            ...card,
-                            providers: [card.provider],
-                        })
-                    } else {
-                        if (!existing.providers.includes(card.provider)) {
-                            existing.providers.push(card.provider)
-                        }
-
-                        existing.bankName ||= card.bankName
-                        existing.bankCountryCode ||= card.bankCountryCode
-                    }
-                }
-
-                const uniqueCardTokens = [...deduplicated.values()]
+                const cardTokens = await fetchCardTokens(userId)
 
                 return {
-                    cardTokens: uniqueCardTokens,
+                    cardTokens: deduplicateCardTokens(cardTokens),
                 }
             },
         },
