+++ apps/condo/domains/acquiring/schema/DeleteCardBindingService.js
@@ -2,14 +2,10 @@
  * Generated by `createservice acquiring.DeleteCardBindingService`
  */
 
-const { getLogger } = require('@open-condo/keystone/logging')
-const { GQLCustomSchema, find } = require('@open-condo/keystone/schema')
+const { GQLCustomSchema } = require('@open-condo/keystone/schema')
 
 const access = require('@condo/domains/acquiring/access/DeleteCardBindingService')
-const { ACQUIRING_INTEGRATION_ONLINE_PROCESSING_TYPE } = require('@condo/domains/acquiring/constants/integration')
-const { deleteUserCard } = require('@condo/domains/acquiring/utils/serverSchema/OnlineInteraction')
-
-const logger = getLogger()
+const { deleteCardBinding } = require('@condo/domains/acquiring/utils/serverSchema/deleteCardBinding/utils')
 
 const DeleteCardBindingService = new GQLCustomSchema('DeleteCardBindingService', {
     types: [
@@ -37,38 +33,14 @@ const DeleteCardBindingService = new GQLCustomSchema('DeleteCardBindingService',
             access: access.canDeleteCardBinding,
             schema: 'deleteCardBinding(data: DeleteCardBindingInput!): DeleteCardBindingOutput!',
             resolver: async (parent, args) => {
-                const { data: { user: { id: userId }, cardId } } = args
+                const {
+                    data: {
+                        user: { id: userId },
+                        cardId,
+                    },
+                } = args
 
-                const acquiringIntegrations = await find('AcquiringIntegration', {
-                    type: ACQUIRING_INTEGRATION_ONLINE_PROCESSING_TYPE,
-                    getUserCardsUrl_not: null,
-                    deleteUserCardUrl_not: null,
-                    deletedAt: null,
-                })
-
-                const results = await Promise.allSettled(
-                    acquiringIntegrations.map(async integration =>
-                        deleteUserCard(
-                            integration.deleteUserCardUrl,
-                            userId,
-                            cardId,
-                            integration.id,
-                        )
-                    )
-                )
-
-                let success = false
-
-                for (const result of results) {
-                    if (result.status === 'fulfilled') {
-                        success = true
-                    } else {
-                        logger.error({
-                            msg: 'Failed to delete card binding',
-                            err: result.reason,
-                        })
-                    }
-                }
+                await deleteCardBinding(userId, cardId)
 
                 return { status: 'ok' }
             },
