+++ apps/condo/domains/acquiring/utils/serverSchema/cardsOnlineInteraction.js
