+++ apps/condo/domains/acquiring/utils/serverSchema/deleteCardBinding/utils.js
@@ -1,0 +1,39 @@
+const { find } = require('@open-condo/keystone/schema')
+
+const { ACQUIRING_INTEGRATION_ONLINE_PROCESSING_TYPE } = require('@condo/domains/acquiring/constants/integration')
+const { deleteUserCard } = require('@condo/domains/acquiring/utils/serverSchema/cardsOnlineInteraction')
+
+async function deleteCardBinding (userId, cardId) {
+    const acquiringIntegrations = await find('AcquiringIntegration', {
+        type: ACQUIRING_INTEGRATION_ONLINE_PROCESSING_TYPE,
+        getUserCardsUrl_not: null,
+        deleteUserCardUrl_not: null,
+        deletedAt: null,
+    })
+
+    const results = await Promise.allSettled(
+        acquiringIntegrations.map(integration =>
+            deleteUserCard(
+                integration.deleteUserCardUrl,
+                userId,
+                cardId,
+                integration.id,
+            )
+        )
+    )
+
+    for (const result of results) {
+        if (result.status === 'fulfilled') {
+            continue
+        }
+
+        logger.error({
+            msg: 'Failed to delete card binding',
+            err: result.reason,
+        })
+    }
+}
+
+module.exports = {
+    deleteCardBinding,
+}
